refactor ComputeInternalForce in prep for semi lagrangian
